requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
python-dotenv>=1.0.0
gunicorn>=22.0.0
eventlet>=0.36.1
gigachat>=0.1.38
